changed quantity datatype from Int to Long

+++ src/main/kotlin/com/vauldex/inventory_management/domain/dto/request/ProductEditRequest.kt
@@ -4,7 +4,7 @@ import com.vauldex.inventory_management.domain.entity.CategoryEntity
 import com.vauldex.inventory_management.domain.entity.ProductEntity
 import java.util.UUID
 
-data class ProductEditRequest(val id: UUID, val productName: String, val quantity: Int, val category: String)
+data class ProductEditRequest(val id: UUID, val productName: String, val quantity: Long, val category: String)
 
 {
     fun toEntity( cat: CategoryEntity): ProductEntity = ProductEntity(
