+++ src/main/kotlin/com/vauldex/inventory_management/domain/dto/response/ProductResponse.kt
@@ -2,4 +2,4 @@ package com.vauldex.inventory_management.domain.dto.response
 
 import java.util.UUID
 
-data class ProductResponse(val productName: String, val quantity: Int, val id: UUID)
+data class ProductResponse(val productName: String, val quantity: Long, val id: UUID)
