+++ src/main/resources/db/migration/V1.1__initial_schema.sql
@@ -44,7 +44,7 @@ CREATE TABLE products(
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     category_id UUID REFERENCES categories(id) ON DELETE NO ACTION,
     product_name VARCHAR(50) NOT NULL,
-    quantity INTEGER NOT NULL,
+    quantity BIGINT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
